requirements.txt
openai==1.3.5
azure-cognitiveservices-speech==1.31.0
streamlit==1.29.0
python-dotenv==1.0.0
sounddevice==0.4.6
soundfile==0.12.1
numpy==1.24.3
SQLAlchemy==2.0.23
alembic==1.13.0
